refactor: componentize activity reducer
