+++ renderer/containers/Activity/Activity.js
@@ -6,7 +6,7 @@ import Activity from 'components/Activity'
 const mapStateToProps = state => ({
   currentActivity: activitySelectors.currentActivity(state)(state),
   isErrorDialogOpen: activitySelectors.isErrorDialogOpen(state),
-  errorDialogDetails: activitySelectors.errorDialogDetailsSelector(state),
+  errorDialogDetails: activitySelectors.errorDialogDetails(state),
 })
 
 const mapDispatchToProps = {
