+++ renderer/reducers/activity/constants.js
@@ -1,0 +1,10 @@
+export const SHOW_ACTIVITY_MODAL = 'SHOW_ACTIVITY_MODAL'
+export const HIDE_ACTIVITY_MODAL = 'HIDE_ACTIVITY_MODAL'
+export const CHANGE_FILTER = 'CHANGE_FILTER'
+export const UPDATE_SEARCH_TEXT = 'UPDATE_SEARCH_TEXT'
+export const FETCH_ACTIVITY_HISTORY = 'FETCH_ACTIVITY_HISTORY'
+export const FETCH_ACTIVITY_HISTORY_SUCCESS = 'FETCH_ACTIVITY_HISTORY_SUCCESS'
+export const FETCH_ACTIVITY_HISTORY_FAILURE = 'FETCH_ACTIVITY_HISTORY_FAILURE'
+export const OPEN_ERROR_DETAILS_DIALOG = 'OPEN_ERROR_DETAILS_DIALOG'
+export const CLOSE_ERROR_DETAILS_DIALOG = 'CLOSE_ERROR_DETAILS_DIALOG'
+export const SET_HAS_NEXT_PAGE = 'SET_HAS_NEXT_PAGE'
