+++ renderer/reducers/activity/index.js
@@ -1,0 +1,6 @@
+import activityReducer from './reducer'
+
+export default activityReducer
+export activitySelectors from './selectors'
+export * from './constants'
+export * from './reducer'
