+++ renderer/reducers/activity/messages.js
@@ -1,0 +1,9 @@
+import { defineMessages } from 'react-intl'
+
+// use pattern [reducer]_message
+
+/* eslint-disable max-len */
+export default defineMessages({
+  activity_invoice_download_error: 'An error has occurred',
+  activity_invoice_download_success: 'Download Complete',
+})
